Update feature scan qr code to redirect view order

+++ src/main/java/com/example/haus/domain/dto/request/promotion/PromotionRequestDto.java
@@ -39,9 +39,9 @@ public class PromotionRequestDto {
     LocalDate endDate;
 
     @PositiveOrNull
-    Float minPriceOrder;
+    Long minPriceOrder;
     @PositiveOrNull
-    Float maxPriceOrder;
+    Long maxPriceOrder;
 
     @NotNull(message = ErrorMessage.Promotion.ERR_PROMOTION_DISCOUNT_PERCENT_NOT_NULL)
     @Min(value = 0, message = ErrorMessage.Promotion.ERR_PROMOTION_DISCOUNT_PERCENT_INVALID)
